pyproject.toml
[project]
name = "llms-from-scratch"
version = "0.1.0"
description = "Training LLM from scratch."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cython>=3.1.4",
    "einops==0.8.1",
    "jupyterlab>=4.4.9",
    "lxml>=6.0.2",
    "matplotlib>=3.10.7",
    "numpy==2.3.1",
    "pandas>=2.3.3",
    "psutil==7.0.0",
    "pyspy>=0.1.1",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "spacy>=3.8.7",
    "tensorboard==2.19.0",
    "termcolor==3.1.0",
    "torch==2.8.0",
    "triton",
    "tqdm==4.67.1",
    "wandb>=0.22.0",
]

[tool.setuptools.packages.find]
where = ["."]
